랜덤 페이지 작업중 카멜케이스 처리 안된 건 조치
