+++ src/main/java/com/koreait/movie/dao/RecommandDao.java
@@ -1,5 +1,16 @@
 package com.koreait.movie.dao;
 
-public interface RecommandDao {
+import java.util.List;
 
+import com.koreait.movie.dto.MovieDto;
+
+public interface RecommandDao {
+	
+	public String movieTitle(int no);
+	
+	public int movieCount();
+	
+	public String movieSitu(int no);
+	
+	public List<MovieDto> movieList(int no);
 }
