+++ src/main/java/com/koreait/movie/dao/mapper/recommandMapper.xml
@@ -4,8 +4,26 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 
 
-<mapper namespace="com.koreait.movie.dao.RecommandDao">
+<mapper namespace="com.koreait.movie.dao.Recommand_dao">
+	
+	<select id="movieCount" resultType="int">
+		select count(*) 
+		 from movie
+	</select>
 	
 	
+	<select id="movieTitle" parameterType = "int" resultType="String">
+		select movie_title
+			from movie
+		where movie_no = #{no}
+	</select>
+	
+	<select id="movieList" parameterType="int" resultType="com.koreait.movie.dto.MovieDto">
+ 		select distinct a.movie_title
+ 			from movie a, genre b, movie_genre, movie_genre c
+ 		where b.genre_no = #{situNo}
+ 			and a.movie_no = c.movie_no
+	</select>
 	
 </mapper>
+
